Allow the :pool_timeout option to be a float
Before, only an integer was allowed.  This is mostly useful for
testing purposes, since I don't want to have to sleep for a
second to test the related code.

+++ lib/sequel/connection_pool/threaded.rb
@@ -30,7 +30,7 @@ class Sequel::ThreadedConnectionPool < Sequel::ConnectionPool
     @connection_handling = opts[:connection_handling]
     @available_connections = []
     @allocated = {}
-    @timeout = Integer(opts[:pool_timeout] || 5)
+    @timeout = Float(opts[:pool_timeout] || 5)
     @sleep_time = Float(opts[:pool_sleep_time] || 0.001)
   end
